fixed 5b4dd24 from https://gitee.com/su-feng6/graphic_standard/pulls/3901 fix webgl codedex

+++ interfaces/kits/napi/graphic/webgl/src/webgl/src/webgl2_rendering_context_base.cpp
@@ -4142,7 +4142,7 @@ napi_value WebGL2RenderingContextBase::InvalidateSubFramebuffer(napi_env env, na
                 return nullptr;
             }
             LOGI("WebGL2 WebGL2RenderingContextBase::invalidateSubFramebuffer ele = %{public}u", ele);
-            attachments1[i] = static_cast<uint64_t>(ele);
+            attachments1[i] = ele;
         }
         attachments = attachments1;
     } else {
